+++ src/option.py
@@ -5,13 +5,15 @@ Module option.py
 Contains the Option class
 """
 
-__version__ = '1.1'
+__version__ = '1.2'
 __author__ = 'Eric G.D'
 
-from typing import List, Tuple
+from random import getrandbits
+from typing import Any, Dict, Iterable, List, Tuple, Union
 
 import numpy as np
 
+from src.constants import TTFlag
 from src.piece import Piece
 
 
@@ -23,16 +25,24 @@ class Option:
     Contains all the data of a single option (move) the computer can make
     """
 
-    def __init__(self, game_state: List[List[Piece]], i: int, j: int, piece: Piece) -> None:
-        self.__game_state = np.array(game_state)
-        self.__index = i, j
-        self.__piece = piece
+    def __init__(self, game_state: np.ndarray, piece: Piece, i: int = -1, j: int = -1) -> None:
+        self.__game_state: GameState = GameState(game_state)
+        self.__piece: Piece = piece
+        self.__index: Tuple[int, int] = (i, j)
 
     def __repr__(self) -> str:
-        return f'Option(\n\tGameState: {self.__game_state}\n\tIndex: {self.__index}\n\tPiece: {self.__piece}\n)\n'
+        return 'Option({0!r}, {1!r}, {2[0]}, {2[1]})'.format(self.__game_state, self.__piece, self.__index)
+
+    def __hash__(self) -> int:
+        return hash(self.__game_state) ^ GameState.selected_table[self.__piece.id]
+
+    def __eq__(self, other: Any) -> bool:
+        return isinstance(other, Option) \
+               and self.__game_state == other.__game_state \
+               and self.__piece == other.piece
 
     @property
-    def game_state(self) -> np.array:
+    def game_state(self) -> 'GameState':
         return self.__game_state
 
     @property
@@ -50,3 +60,101 @@ class Option:
     @property
     def piece(self) -> Piece:
         return self.__piece
+
+
+class Transposition:
+
+    def __init__(self, value: int, flag: TTFlag, depth: int) -> None:
+        self.__value: int = value
+        self.__flag: TTFlag = flag
+        self.__depth: int = depth
+
+    def __repr__(self) -> str:
+        return f'Transposition({self.__value}, {str(self.__flag)}, {self.__depth})'
+
+    @staticmethod
+    def get_flag(value: int, alpha: float, beta: float) -> TTFlag:
+        """
+        :param value:   The value of the current node being calculated
+        :param alpha:   The original alpha value before the search began
+        :param beta:    The latest beta value of the search
+        :return:        The transposition flag for the current search
+        """
+        assert alpha <= beta
+        if value <= alpha:
+            return TTFlag.upper_bound
+        if value >= beta:
+            return TTFlag.lower_bound
+        return TTFlag.exact
+
+    @property
+    def value(self) -> int:
+        return self.__value
+
+    @property
+    def flag(self) -> TTFlag:
+        return self.__flag
+
+    @property
+    def depth(self) -> int:
+        return self.__depth
+
+
+class GameState:
+    LENGTH: int = 4
+    zobrist_table: Union[None, List[List[List[int]]]] = None
+    selected_table: Union[None, List[int]] = None
+    transposition_table: Union[None, Dict[Option, Transposition]] = None
+
+    def __init__(self, state: Union[np.ndarray, List[List[Piece]]]):
+        if not len(state) == len(state[0]) == GameState.LENGTH:
+            raise ValueError("Invalid dimensions for :state:, should be a {0}x{0} array!".format(GameState.LENGTH))
+        self.__board: np.ndarray = np.array(state)
+        if GameState.zobrist_table is None:
+            GameState.__zobrist_init()
+        if GameState.transposition_table is None:
+            GameState.transposition_table = {}
+
+    @property
+    def board(self) -> np.ndarray:
+        return self.__board
+
+    def __getitem__(self, key: Any) -> np.ndarray:
+        return self.__board[key]
+
+    def __iter__(self) -> Iterable[np.ndarray]:
+        return iter(self.board)
+
+    def __hash__(self) -> int:
+        """
+        Uses the Zobrist hashing algorithm to provide a unique hash for the current game state
+        ..seealso:  https://en.wikipedia.org/wiki/Zobrist_hashing
+        :return:    A unique hash value for the current game state
+        """
+        hash_ = 0
+        for i in range(len(self.board)):
+            for j in range(len(self.board[i])):
+                if self.board[i][j] is None:
+                    continue
+                hash_ ^= GameState.zobrist_table[i][j][self.board[i][j].id]
+        return hash_
+
+    def __eq__(self, other: Any) -> bool:
+        return isinstance(other, GameState) and np.array_equal(self.board, other.board)
+
+    def __len__(self) -> int:
+        return len(self.board)
+
+    def __repr__(self) -> str:
+        return 'Gamestate({0!r})'.format(self.__board.tolist())
+
+    @staticmethod
+    def __zobrist_init() -> None:
+        """
+        Initialises the zobrist table with random 64 bit numbers
+        :return:    None
+        """
+        table = [[[getrandbits(64) for _ in range(Piece.MAX_NUM + 1)]  # For pieces in the board
+                  for _ in range(GameState.LENGTH)] for _ in range(GameState.LENGTH)]
+        GameState.zobrist_table = table
+        GameState.selected_table = [getrandbits(64) for _ in range(Piece.MAX_NUM + 1)]  # For selected piece